requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
packaging==25.0
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
